+++ cella-lints/build.dart
@@ -10,13 +10,13 @@ void main() {
   final monorepo = dirname(root);
 
   // Copy license file into current child repo.
-  final licenseFilename = 'license'.toUpperCase();
-  File(join(monorepo, licenseFilename)).copySync(join(root, licenseFilename));
+  final license = 'license'.toUpperCase();
+  File(join(monorepo, license)).copySync(join(root, license));
 
   // Sync contributors into current child repo.
-  final repoName = basename(root);
-  final contributorsFilename = '${'contributors'.toUpperCase()}.yaml';
-  final path = join(monorepo, contributorsFilename);
+  final name = basename(root);
+  final contributors = '${'contributors'.toUpperCase()}.yaml';
+  final path = join(monorepo, contributors);
   final raw = loadYaml(File(path).readAsStringSync());
   if (raw is! Map) throw Exception('invalid structure: $path');
   final handler = <String, dynamic>{};
@@ -25,20 +25,18 @@ void main() {
     if (username is! String) continue;
     final contributor = raw[username];
     if (contributor is! Map) continue;
-    final repos = contributor['repo'];
-    if (repos is! List<dynamic>) continue;
-    if (repos.contains(repoName)) {
+    final repo = contributor['repo'];
+    if (repo == name || (repo is List<dynamic> && repo.contains(name))) {
       handler[username] = {...contributor}..remove('repo');
     }
   }
   final editor = YamlEditor('')..update([], handler);
-  File(join(root, contributorsFilename)).writeAsStringSync('$editor\n');
+  File(join(root, contributors)).writeAsStringSync('$editor\n');
 
   // Auto generate .pubignore.
   final content = [
-    contributorsFilename.toLowerCase(),
-    licenseFilename.toLowerCase(),
-    'build.dart',
+    contributors.toLowerCase(),
+    license.toLowerCase(),
   ].map((item) => '!$item').join('\n');
   File(join(root, '.pubignore')).writeAsStringSync('$content\n');
 }